refactor: move evaluation functions to scheduler package
