+++ internal/digraph/scheduler/eval_test.go
@@ -1,20 +1,22 @@
-package executor
+package scheduler_test
 
 import (
 	"context"
 	"testing"
 
 	"github.com/dagu-org/dagu/internal/digraph"
+	"github.com/dagu-org/dagu/internal/digraph/executor"
+	"github.com/dagu-org/dagu/internal/digraph/scheduler"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestEvalString(t *testing.T) {
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := NewEnv(ctx, digraph.Step{Name: "test-step"})
+	env := executor.NewEnv(ctx, digraph.Step{Name: "test-step"})
 	env.Variables.Store("TEST_VAR", "TEST_VAR=hello")
 	env.Envs["ANOTHER_VAR"] = "world"
-	ctx = WithEnv(ctx, env)
+	ctx = executor.WithEnv(ctx, env)
 
 	tests := []struct {
 		name     string
@@ -50,7 +52,7 @@ func TestEvalString(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			result, err := EvalString(ctx, tt.input)
+			result, err := scheduler.EvalString(ctx, tt.input)
 			assert.NoError(t, err)
 			assert.Equal(t, tt.expected, result)
 		})
@@ -60,13 +62,13 @@ func TestEvalString(t *testing.T) {
 func TestEvalBool(t *testing.T) {
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := NewEnv(ctx, digraph.Step{Name: "test-step"})
+	env := executor.NewEnv(ctx, digraph.Step{Name: "test-step"})
 	env.Variables.Store("TRUE_VAR", "TRUE_VAR=true")
 	env.Variables.Store("FALSE_VAR", "FALSE_VAR=false")
 	env.Variables.Store("ONE_VAR", "ONE_VAR=1")
 	env.Variables.Store("ZERO_VAR", "ZERO_VAR=0")
 	env.Variables.Store("INVALID_VAR", "INVALID_VAR=not-a-bool")
-	ctx = WithEnv(ctx, env)
+	ctx = executor.WithEnv(ctx, env)
 
 	tests := []struct {
 		name     string
@@ -126,7 +128,7 @@ func TestEvalBool(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			result, err := EvalBool(ctx, tt.input)
+			result, err := scheduler.EvalBool(ctx, tt.input)
 			if tt.wantErr {
 				assert.Error(t, err)
 			} else {
@@ -152,11 +154,11 @@ type NestedStruct struct {
 func TestEvalObject(t *testing.T) {
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := NewEnv(ctx, digraph.Step{Name: "test-step"})
+	env := executor.NewEnv(ctx, digraph.Step{Name: "test-step"})
 	env.Variables.Store("NAME_VAR", "NAME_VAR=John")
 	env.Variables.Store("DESC_VAR", "DESC_VAR=Developer")
 	env.Variables.Store("NESTED_VAR", "NESTED_VAR=NestedValue")
-	ctx = WithEnv(ctx, env)
+	ctx = executor.WithEnv(ctx, env)
 
 	// Create a test struct
 	testObj := TestStruct{
@@ -179,12 +181,12 @@ func TestEvalObject(t *testing.T) {
 	}
 
 	// Test EvalObject
-	result, err := EvalObject(ctx, testObj)
+	result, err := scheduler.EvalObject(ctx, testObj)
 	assert.NoError(t, err)
 	assert.Equal(t, expected, result)
 
 	// Test with a non-struct type
-	_, err = EvalObject(ctx, "not a struct")
+	_, err = scheduler.EvalObject(ctx, "not a struct")
 	assert.NoError(t, err)
 }
 
@@ -192,11 +194,11 @@ func TestEvalObject(t *testing.T) {
 func TestEvalObjectWithExecutorConfig(t *testing.T) {
 	// Create a test context with environment variables
 	ctx := context.Background()
-	env := NewEnv(ctx, digraph.Step{Name: "test-step"})
+	env := executor.NewEnv(ctx, digraph.Step{Name: "test-step"})
 	env.Variables.Store("EXECUTOR_TYPE", "EXECUTOR_TYPE=docker")
 	env.Variables.Store("HOST_VAR", "HOST_VAR=localhost")
 	env.Variables.Store("PORT_VAR", "PORT_VAR=8080")
-	ctx = WithEnv(ctx, env)
+	ctx = executor.WithEnv(ctx, env)
 
 	// Create an ExecutorConfig with variables
 	config := digraph.ExecutorConfig{
@@ -223,7 +225,7 @@ func TestEvalObjectWithExecutorConfig(t *testing.T) {
 	}
 
 	// Test EvalObject
-	result, err := EvalObject(ctx, config.Config)
+	result, err := scheduler.EvalObject(ctx, config.Config)
 	assert.NoError(t, err)
 
 	// Check Config map values
